[CMake] Bump minimum required version to 2.8.3. https://bugs.webkit.org/show_bug.cgi?id=93450
Reviewed by Daniel Bates.

Following the discussion in bug 93189, it was decided that bumping
the minimum required version from 2.8.0 to 2.8.3 was a good thing
due to additional features, such as improved
FIND_PACKAGE_HANDLE_STANDARD_ARGS() support and better argument
handling in the PKG_CHECK_MODULES() call.

CMake 2.8.3 was released almost two years ago, so people are
expected to have it or a later version installed.

* CMakeLists.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125228 268f45cc-cd09-0410-ab3c-d52691b4dbfc