itemType.add should treat \t as a space. https://bugs.webkit.org/show_bug.cgi?id=92991
Reviewed by Ryosuke Niwa.

Source/WebCore: 

Earlier we used to append a space character i.e, ' ' to the string builder
if the last character of input string is not a space character in addToken()
method. We should add an space character if the last character of input is
not a HTML Space character like, '\n', '\r', '\t', '\f' or ' '.

We can observe the same behavior in FireFox, Opera and IE. They also treat
HTML space character as a space.

Test: fast/dom/MicroData/domsettabletokenlist-attributes-add-token.html

* html/DOMTokenList.cpp:
(WebCore::DOMTokenList::addToken):

LayoutTests: 

Added test to verifiy itemtype, itemref, itemprop attributes behavior on adding
a token in presence of trailing space characters.
Rebased existing test cases.

* fast/dom/HTMLElement/class-list-expected.txt:
* fast/dom/HTMLElement/class-list-quirks-expected.txt:
* fast/dom/HTMLElement/script-tests/class-list.js:
* fast/dom/HTMLOutputElement/dom-settable-token-list-expected.txt:
* fast/dom/HTMLOutputElement/script-tests/dom-settable-token-list.js:
* fast/dom/MicroData/domsettabletokenlist-attributes-add-token-expected.txt: Added.
* fast/dom/MicroData/domsettabletokenlist-attributes-add-token.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125257 268f45cc-cd09-0410-ab3c-d52691b4dbfc