[BlackBerry] Some media controls are mispositioned for dynamic live streams (HLS) https://bugs.webkit.org/show_bug.cgi?id=94176
Patch by Max Feil <mfeil@rim.com> on 2012-08-16
Reviewed by Antonio Gomes.

.:

An automated layout test is not possible for this patch because
dynamic live streams require a special dedicated web server.
Putting an external video URL into an automated test is not
correct either. So I have created a manual test that points to
an external HLS video that works today.

* ManualTests/blackberry/video-hls-controls.html: Added.

Source/WebCore:

The default HTML5 media controls for dynamic live streams have
no timeline or timeline container, which for BlackBerry results
in mispositioning of the buttons that are supposed to be to
the right of the timeline (fullscreen and mute). Instead of
being right justified they incorrectly appear on the left next
to the play button. The fix is to explicitly position these 2
buttons whenever the media duration is infinite (indicating a
live stream).

Manual test: ManualTests/blackberry/video-hls-controls.html

* platform/blackberry/RenderThemeBlackBerry.cpp:
(WebCore::RenderThemeBlackBerry::adjustMediaControlStyle):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125811 268f45cc-cd09-0410-ab3c-d52691b4dbfc