IndexedDB: Test if getting a TRANSACTION_INACTIVE_ERR is possible when setting version https://bugs.webkit.org/show_bug.cgi?id=94269
Reviewed by Ojan Vafai.

I don't see a path through which this code is exercised.
ASSERT_NOT_REACHED doesn't cause any of our tests to fail. It might be
timing-related, in which case this ASSERT should be sporadically hit.
If this ASSERT is hit, see if we can write a test that exercises it or
at least add comments explaining how it gets exercised. If this ASSERT
is not hit after a while, change it to
ASSERT(transaction.scheduleTask(...)).

No new tests, we're looking for existing tests to fail.

* Modules/indexeddb/IDBDatabaseBackendImpl.cpp:
(WebCore::IDBDatabaseBackendImpl::setVersion):
(WebCore::IDBDatabaseBackendImpl::runIntVersionChangeTransaction):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125943 268f45cc-cd09-0410-ab3c-d52691b4dbfc