Track -webkit property usage. https://bugs.webkit.org/show_bug.cgi?id=93420
Patch by Tab Atkins <tabatkins@google.com> on 2012-08-21
Reviewed by Ojan Vafai.

First draft of an attempt to track all usage of -webkit prefixed properties across the web.
This attempt is dumb, but should provide useful data as a first-pass.
I plan to optimize this for better data collection in the future.

No tests added, as this is untestable currently.
It should have zero effect besides histogramming.

* css/CSSParser.cpp:
(WebCore::cssPropertyID):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126201 268f45cc-cd09-0410-ab3c-d52691b4dbfc