[chromium] Have RenderSurface create and add its own generated RenderPass https://bugs.webkit.org/show_bug.cgi?id=94957
Reviewed by Adrienne Walker.

Source/WebCore:

Have CCRenderSurface create and add its RenderPass to the appropriate
data structures. Previously CCLayerTreeHostImpl had this responsibility
but this limits us to a single RenderPass per RenderSurface. In ubercomp
we have layers that generate many RenderPasses that all contribute to a
single RenderSurface. This target RenderSurface can create all the needed
RenderPasses in the correct order for itself when asked to now.

We invent CCRenderPassSink and make FrameData implement it, so that the
RenderSurface does not need to know the details of how the RenderPasses
it generates will be stored.

Test: CCRenderSurfaceTest.sanityCheckSurfaceCreatesCorrectRenderPass

* WebCore.gypi:
* platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp:
(WebCore::CCLayerTreeHostImpl::FrameData::appendRenderPass):
(WebCore):
(WebCore::CCLayerTreeHostImpl::calculateRenderPasses):
* platform/graphics/chromium/cc/CCLayerTreeHostImpl.h:
(FrameData):
* platform/graphics/chromium/cc/CCRenderPassSink.h: Added.
(WebCore):
(CCRenderPassSink):
* platform/graphics/chromium/cc/CCRenderSurface.cpp:
(WebCore::CCRenderSurface::appendRenderPasses):
(WebCore):
* platform/graphics/chromium/cc/CCRenderSurface.h:
(WebCore):
(CCRenderSurface):

Source/WebKit/chromium:

Test that a RenderSurface generates the RenderPass that we expect.

* tests/CCRenderSurfaceTest.cpp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126790 268f45cc-cd09-0410-ab3c-d52691b4dbfc