Stylechecker warns about comparison to zero when comparing to 0.5 https://bugs.webkit.org/show_bug.cgi?id=94913
Reviewed by Dirk Pranke.

According to the python documentation, \W is the character group with all
non-alphanumeric characters, equivalent to [^a-zA-Z0-9_], which is equal to
to [^\w]. We need to exclude the dot "." as well, so that floating point
values do not trigger this warning. Adding a unit test that shows the problem.

* Scripts/webkitpy/style/checkers/cpp.py:
(check_for_comparisons_to_zero): Modifying the regex to not get triggered by comparing to floats.
* Scripts/webkitpy/style/checkers/cpp_unittest.py:
(WebKitStyleTest.test_null_false_zero): Unit test exposing this issue.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127036 268f45cc-cd09-0410-ab3c-d52691b4dbfc