IndexedDB: IDBRequest leaks if IDBCursor closes and no further events fired https://bugs.webkit.org/show_bug.cgi?id=95777
Reviewed by Tony Chang.

IDBRequests are kept alive as long as they may fire events, which includes
if the associated IDBCursor object can be advanced. When the transaction
is finished the IDBCursor is notified which in turn tells the IDBRequest
that the cursor won't be the source of more events. However, if this occurs
and no further events fire the IDBRequest doesn't clear it's "has pending
activity flag". Clear the flag on the notification if the request is
otherwise complete.

No new tests - ActiveDOMObjects and leaks are persnickety.

* Modules/indexeddb/IDBRequest.cpp:
(WebCore::IDBRequest::finishCursor):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127518 268f45cc-cd09-0410-ab3c-d52691b4dbfc