CSS3 calc: expressions with 'em' units do not zoom correctly. https://bugs.webkit.org/show_bug.cgi?id=95705
Reviewed by Ojan Vafai.

Source/WebCore:

Each primitive value in a CSS calc expression now has (zoom) multiplier and scale factor applied
independently. Previously the multiplier and a single scale factor was applied to the expression
as a whole, but this failed to account for expressions involving font relative units. This is
because the multiplier should not be applied to font relative units.

Test: css3/calc/zoom-with-em.html

* css/CSSPrimitiveValue.cpp:
(WebCore::CSSPrimitiveValue::computeLengthDouble):

LayoutTests:

* css3/calc/zoom-with-em-expected.txt: Added.
* css3/calc/zoom-with-em.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127557 268f45cc-cd09-0410-ab3c-d52691b4dbfc