[BlackBerry] Rotation and then reload causes less sized webpage (PR 190469) https://bugs.webkit.org/show_bug.cgi?id=95870
Reviewed by George Staikos.
Reviewed internally by Jacky Jiang.

Use the new zoom to fit scale if the page was zoomed to fit initial scale
When a page is reloaded the layout size may be changed. So the contents
size on which the initial scale depends may be changed. Use the new initial
scale in this case to make sure the page will be zoomed to correct scale.

* WebCoreSupport/FrameLoaderClientBlackBerry.cpp:
(WebCore::FrameLoaderClientBlackBerry::restoreViewState):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127628 268f45cc-cd09-0410-ab3c-d52691b4dbfc