16 bit JSRopeString up converts an 8 bit fibers to 16 bits during resolution https://bugs.webkit.org/show_bug.cgi?id=95810
Reviewed by Benjamin Poulain.

Source/JavaScriptCore: 

Added 8 bit path that copies the contents of an 8 bit fiber to the 16 bit buffer
when resolving a 16 bit rope.

* runtime/JSString.cpp:
(JSC::JSRopeString::resolveRopeSlowCase):

Source/WTF: 

New copy routine that takes an 8 bit source and a 16 bit destination.  Used when copying
the contents of an 8 bit fiber to the 16 bit buffer when resolving a 16 bit rope.

* wtf/text/StringImpl.h:
(WTF::StringImpl::copyChars):
(StringImpl):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127809 268f45cc-cd09-0410-ab3c-d52691b4dbfc