[GTK] Move user agent helpers to WebCore https://bugs.webkit.org/show_bug.cgi?id=95745
Reviewed by Carlos Garcia Campos.

.:

Added an autoconf step for the new user agent shared code header in
WebCore. This is necessary so that we can use the user agent version
based on the information in the configure.ac file.

* configure.ac:

Source/WebCore:

Moved the user agent determination code from WebKit1 to WebCore so that it can
be shared with WebKit2. Also add a method for creating the user agent given
a non-default application name.

* GNUmakefile.list.am: Add new files to the list.
* platform/gtk/UserAgentGtk.cpp: Added.
* platform/gtk/UserAgentGtk.h.in: Added.

Source/WebKit/gtk:

Move the code for determining the user agent to WebCore and have WebKit1
use the new shared code.

* webkit/webkitwebsettings.cpp: Use the WebCore code to determine the user
agent in WebKit1.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@127889 268f45cc-cd09-0410-ab3c-d52691b4dbfc