[Qt] Fix generation of forward headers for generated files in WebCore on Windows https://bugs.webkit.org/show_bug.cgi?id=96245
Patch by Simon Hausmann <simon.hausmann@nokia.com> on 2012-09-10
Reviewed by Tor Arne Vestbø.

We generate the forwarding header files of generated WebCore sources by using
mkdir to ensure the target directory exists, followed by
echo #include "..." > file.h. On Windows mkdir returns with an error if the
target directory exists. For us that is not an error, so work around it with
(mkdir path 2>nul || echo>nul)

* DerivedSources.pri:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128045 268f45cc-cd09-0410-ab3c-d52691b4dbfc