[Qt] Fix passing of defines from the build system
We explicitly have to pass -DENABLE_FOO=0 for every single feature
that's not explicitly enabled in WEBKIT_CONFIG, since Platform.h
will add it's own defaults if the features are not defined.

At some point we might want to run a configure-test to inspect what
Platform.h will do, so we can sync up the build system's view of
the feature-defines with the compile-time situation.

Reviewed by Simon Hausmann.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128179 268f45cc-cd09-0410-ab3c-d52691b4dbfc