LocaleMac should fallback to defaultLanguage for an invalid locale https://bugs.webkit.org/show_bug.cgi?id=96744
Reviewed by Kent Tamura.

Source/WebCore:

There is no way to tell if a locale identifier is valid or not in Cocoa
so we look if we have a matching language in the list of
ISOLanguageCodes given by NSLocale.

Covered in LocaleMacTest.invalidLocale.

* platform/text/mac/LocaleMac.mm:
(WebCore):
(WebCore::LocaleMac::LocaleMac):
(WebCore::LocaleMac::create):

Source/WebKit/chromium:

* tests/LocaleMacTest.cpp:
(TEST_F):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128582 268f45cc-cd09-0410-ab3c-d52691b4dbfc