[chromium] Add the ubercomp WebDelegatedRendererLayer https://bugs.webkit.org/show_bug.cgi?id=94145
Reviewed by Adrienne Walker.

Expose a WebDelegatedRendererLayer to allow the browser compositor to
embed such a layer in its tree. This layer will be connected to IPC
machinery in the compositor that is yet TBD which will give a set of
RenderPasses with DrawQuads to the impl copy of the layer.

* Platform.gypi:
* chromium/public/WebCompositorSupport.h:
(WebKit):
(WebKit::WebCompositorSupport::createDelegatedRendererLayer):
(WebCompositorSupport):
* chromium/public/WebDelegatedRendererLayer.h: Added.
(WebKit):
(WebDelegatedRendererLayer):
(WebKit::WebDelegatedRendererLayer::~WebDelegatedRendererLayer):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128630 268f45cc-cd09-0410-ab3c-d52691b4dbfc