[Chromium] Switch back to a fixed fifo path for Android https://bugs.webkit.org/show_bug.cgi?id=97230
Reviewed by Tony Chang.

Because not all external storage cards will be formated using a file
system that supports named pipes, Chromium has been changed to creating
the pipes in a temporary folder on the internal storage. Adapt this in
WebKit so we can continue to run layout tests.

This also includes an *unreviewed* fix for a breakage in the webkitpy
tests I made in r129221. Two lines and related to this code, so I decided
to include it in this change.

* Scripts/webkitpy/layout_tests/port/chromium_android.py:
(ChromiumAndroidDriver.__init__):
(ChromiumAndroidDriver._setup_test):
(ChromiumAndroidDriver._update_version):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129226 268f45cc-cd09-0410-ab3c-d52691b4dbfc