[CSSRegions]Add test for accelerating compositing of elements in RenderFlowThread https://bugs.webkit.org/show_bug.cgi?id=97387
Reviewed by Dirk Schulze.

At this moment, accelerated compositing is disabled for elements inside a RenderFlowThread. This patch adds
a specific test in compositing/regions that is expected to fail until https://bugs.webkit.org/show_bug.cgi?id=84900
is fixed. The code for the new test is taken out of the existing fast/regions/webkit-flow-renderer-layer.html.

* compositing/regions/webkit-flow-renderer-layer-compositing-expected.html: Added.
* compositing/regions/webkit-flow-renderer-layer-compositing.html: Added.
* fast/regions/webkit-flow-renderer-layer-expected.html:
* fast/regions/webkit-flow-renderer-layer.html:
* platform/chromium/TestExpectations:
* platform/efl/TestExpectations:
* platform/gtk/TestExpectations:
* platform/mac/TestExpectations:
* platform/qt/Skipped:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129304 268f45cc-cd09-0410-ab3c-d52691b4dbfc