[GTK] Add Select All method to WebKit2 GTK+ API https://bugs.webkit.org/show_bug.cgi?id=97460
Patch by Simon Pena <spena@igalia.com> on 2012-09-24
Reviewed by Martin Robinson.

Following the same approach used when added Cut, Copy and Paste,
the Select All method is added to the WebKit2 GTK+ API.

This introduces a new macro in the WebKitEditingCommands,
updates the documentation, and includes a new unit test.

* UIProcess/API/gtk/WebKitEditingCommands.h: Add a new macro for
the Select All command.
* UIProcess/API/gtk/docs/webkit2gtk-sections.txt: Include the Select All
command in the documentation.
* UIProcess/API/gtk/tests/TestWebViewEditor.cpp: Cover the new command
with a unit test.
(testWebViewEditorSelectAll):
(beforeAll):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129399 268f45cc-cd09-0410-ab3c-d52691b4dbfc