Allow any kind of v8::Handle in invokeCallback instead of requiring a v8::Persistent. https://bugs.webkit.org/show_bug.cgi?id=97956
Patch by Elliott Sprehn <esprehn@chromium.org> on 2012-09-28
Reviewed by Adam Barth.

Use v8::Handle instead of v8::Persistent for the callback argument on invokeCallback.
There doesn't seem to be any reason for requiring a Persistent, and it makes it harder
to use this API.

This is factored out of http://wkbug.com/93661

No new tests, this is functionally equivalent.

* bindings/v8/V8Callback.cpp:
(WebCore::invokeCallback):
* bindings/v8/V8Callback.h:
(WebCore):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129970 268f45cc-cd09-0410-ab3c-d52691b4dbfc