Enable SATURATED_LAYOUT_ARITHMETIC for chromium https://bugs.webkit.org/show_bug.cgi?id=95053
Reviewed by Abhishek Arya.

Enable the SATURATED_LAYOUT_ARITHMETIC flag for the chromium port.
This changes the behavior of FractionalLayoutUnit to clamp to the
max or min value instead of overflowing.

This may very well impact performance so the current plan is to enable
it for a couple of hours to a day to collect performance data and then
disable it again until we've had a chance to review the perf data.

* wtf/Platform.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129982 268f45cc-cd09-0410-ab3c-d52691b4dbfc