SVGAttributeHashTranslator does not need to copy QualifiedName in the common case https://bugs.webkit.org/show_bug.cgi?id=98473
Reviewed by Adam Barth.

I tested this using instruments on a test case which modified SVG attributes in a loop.
I believe pdr has some perf-tests in this area, but they weren't needed here.  A simple sample showed this as a huge win,
since we're no longer creating a QualifiedName (and thus adding it to the hash) on each QualifiedName-based lookup in SVG.

* svg/SVGElement.h:
(WebCore::SVGAttributeHashTranslator::hash):
(WebCore::SVGAttributeHashTranslator::equal):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130456 268f45cc-cd09-0410-ab3c-d52691b4dbfc