Changed debug-safari, debug-minibrowser and debug-test-runner to use LLDB by default when using Xcode 4.5 or later.
Reviewed by Anders Carlsson.

* Scripts/webkitdirs.pm:
(determineDebugger): Changed the default debugger from GDB to LLDB when the Xcode version
is 4.5 or later. Added a --use-gdb switch for overriding this.
(printHelpAndExitForRunAndDebugWebKitAppIfNeeded): Updated to mention --use-gdb and the
how the default debugger choice depends on the Xcode version. Also changed the description
of --no-saved-state to use the same terminology Xcode’s scheme editor uses to describe this
option, and clarified to which OS X versions it applies.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130924 268f45cc-cd09-0410-ab3c-d52691b4dbfc