[WebGL] [Mac] only the mapped symbol for the first element of a uniform/attribute array is stored. https://bugs.webkit.org/show_bug.cgi?id=98966
Reviewed by Dean Jackson.

We currently do not add name/mappedName symbol pairs for each element of a uniform/attribute array.
This is because we only add a number of symbols equal to how many symbols the shader compiler tells us there are.
The shader compiler treats an array as a single uniform. We need to be adding in symbols to the map for each element of the array ourselves.
The mappedName of each array element is the same between elements except for the "[index]" at the end so this is easy to do.

Tested using Khronos WebGL conformance suite:
conformance/glsl/misc/glsl-long-variable-names.html

* platform/graphics/ANGLEWebKitBridge.cpp:
(WebCore::getSymbolInfo):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131105 268f45cc-cd09-0410-ab3c-d52691b4dbfc