Expose the main frame's document to injected bundles via the Objective-C API https://bugs.webkit.org/show_bug.cgi?id=99133
Patch by Sam Weinig.

Reviewed by Anders Carlsson.

Allow a bundle to declare a principle class which will be instantiated
by WebKit2 and passed an object to which WKBundleClient calls are
forwarded and which can access the main frame's WKDOMDocument.

* WebKit2.xcodeproj/project.pbxproj:
* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugIn.h:
* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugIn.mm: Instantiate
the bundle's principle class forward WKBundleClient calls to it.
* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugInBrowserContextController.h:
* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugInBrowserContextController.mm:
Add a property called mainFrameDocument which returns a WKDOMDocument.
* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugInBrowserContextControllerInternal.h:
* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugInInternal.h:
* WebProcess/InjectedBundle/mac/InjectedBundleMac.mm:
(WebKit::InjectedBundle::load): Initialize the shared
WKWebProcessPlugInController with the bundle's principle class.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131138 268f45cc-cd09-0410-ab3c-d52691b4dbfc