REGRESSION: Rapid memory growth calling DOM APIs with large strings. <http://webkit.org/b/98498> <rdar://problem/12443926>
Reviewed by Geöff Gären and Änders Cärlssön.

Prevent the selector query cache from growing indefinitely by setting a relaxed limit of 256 entries.
If the cache fills up, remove a random entry before inserting a new one.

While this is unlikely to be a problem on real websites, we definitely shouldn't be adding boundless
caches to WebKit.

* dom/SelectorQuery.cpp:
(WebCore::SelectorQueryCache::add):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131209 268f45cc-cd09-0410-ab3c-d52691b4dbfc