[BlackBerry] Fix WebSockets which was disabled by recent string changes https://bugs.webkit.org/show_bug.cgi?id=99248
Patch by Joe Mason <jmason@rim.com> on 2012-10-15
Reviewed by George Staikos.

PR 223681: The signature of FilterStream::notifyStatusReceived changed to take a
BlackBerry::Platform::String instead of a char *. But SocketStreamHandleBlackBerry was not
updated to the new signature, so the status message is not getting passed on because it is
not calling the correct override.

No new tests because this is fixing a regression.

* platform/network/blackberry/SocketStreamHandle.h:
(SocketStreamHandle):
* platform/network/blackberry/SocketStreamHandleBlackBerry.cpp:
(WebCore::SocketStreamHandle::notifyStatusReceived):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131324 268f45cc-cd09-0410-ab3c-d52691b4dbfc