Remove page visibility hidden histograms https://bugs.webkit.org/show_bug.cgi?id=99360
Reviewed by Ojan Vafai.

The 2 page visibility hidden histograms were added in http://webkit.org/b/85537 to
investigate a potential optimization where we would destroy the render tree when
a tab was hidden. The optimization was abandoned but the histograms were never
removed. They are too narrow to be reused so this change removes them.

* page/Page.cpp:
(WebCore::Page::setVisibilityState):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131391 268f45cc-cd09-0410-ab3c-d52691b4dbfc