[Chromium] The default target for Android builds should be "all" instead of "All" https://bugs.webkit.org/show_bug.cgi?id=98879
Reviewed by Adam Barth.

The default gyp target for Android builds is being overridden in Chromium's
envsetup_functions.sh to "All" (whereas the rest of Chromium and WebKit use "all"),
because Android wants to use a different list of main targets. There is
no need to do this for the WebKit project, and it's breaking ninja.

* gyp_webkit:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131394 268f45cc-cd09-0410-ab3c-d52691b4dbfc