The HTML5 formtarget/formaction/formenctyp/formmethod/formnovalidate attributes don't work in button tags with nested elements https://bugs.webkit.org/show_bug.cgi?id=90539
Patch by Kunihiko Sakamoto <ksakamoto@chromium.org> on 2012-10-17
Reviewed by Kent Tamura.

Source/WebCore:

Fix a bug where form(action|enctype|method|target) attributes of <button> are not
honored if the target of the click event is an element nested within the button.

Test: fast/forms/formaction-attribute.html

* loader/FormSubmission.cpp:
(WebCore::FormSubmission::create): Looks for the attributes for form submission
from the nearest FormControlElement ancestor of the event target.

LayoutTests:

Added a test case to verify that formaction attribute of <button> element is
used when its child <span> is clicked.

* fast/forms/formaction-attribute-expected.txt:
* fast/forms/formaction-attribute.html: Removed unused doneAction2 variable.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131555 268f45cc-cd09-0410-ab3c-d52691b4dbfc