Content of replaced elements should be trimmed to the content edge curve. https://bugs.webkit.org/show_bug.cgi?id=63899
Patch by Douglas Stockwell <dstockwell@chromium.org> on 2012-10-17
Reviewed by Simon Fraser.

Source/WebCore:

When a border-radius is specified the content of replaced elements needs to be clipped to avoid
being painted over the padding or border in the corners. Previously the clip that was applied
was set to the border-edge curve, this patch contracts the clip by the size of the border and
padding (the content-edge curve) to match the css3 spec.
Spec: http://www.w3.org/TR/css3-background/#corner-clipping

Test: fast/replaced/border-radius-clip-content-edge.html

* rendering/RenderBoxModelObject.h:
(RenderBoxModelObject):
* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::paint):
* rendering/RenderWidget.cpp:
(WebCore::RenderWidget::paint):

LayoutTests:

* fast/replaced/border-radius-clip-content-edge-expected.txt: Added.
* fast/replaced/border-radius-clip-content-edge.html: Added.
* platform/fast/replaced/border-radius-clip-content-edge-expected.png: Added.
* platform/chromium/TestExpectations:
* platform/chromium-linux/fast/replaced/border-radius-clip-content-edge-expected.png: Added.
* platform/chromium-linux/ietestcenter/css3/bordersbackgrounds/border-radius-content-edge-001-expected.png:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131557 268f45cc-cd09-0410-ab3c-d52691b4dbfc