Fix the build with a newer version of clang.
Reviewed by Dan Bernstein.

Source/WebKit/mac:

Some of the methods in WebCoreStatistics appear to have been added by someone not familiar with Objective-C.
While it's technically valid to have empty components in a selector, it's rarely done and is not consistent
with our style guidelines. In this particular case it's also done in such a manner that it's ambiguous and
therefore generates a warning in newer versions of the compiler.

Fixes <rdar://problem/12503709>.

* Misc/WebCoreStatistics.h: Remove unused methods and rename the two poorly-named methods that remain.
* Misc/WebCoreStatistics.mm: Ditto.
(-[WebFrame numberOfPagesWithPageWidth:pageHeight:]):
(-[WebFrame printToCGContext:pageWidth:pageHeight:]):

Tools:

Update to accommodate the renamed methods in WebCoreStatistics.

* DumpRenderTree/mac/PixelDumpSupportMac.mm:
(createPagedBitmapContext):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131564 268f45cc-cd09-0410-ab3c-d52691b4dbfc