[EFL] Add background color to the checkbox theme. https://bugs.webkit.org/show_bug.cgi?id=98788
Patch by Byungwoo Lee <bw80.lee@samsung.com> on 2012-10-17
Reviewed by Gyuyoung Kim.

Background color is added to the checkbox theme for identifying
checkbox status easily.

Checkbox theme uses images which has transparent background, and the
theme itself doesn't have the background color.

When the background of checkbox is black, checked/unchecked status
cannot be distinguished because check mark color is black.

And it is hard to identify whether the checkbox is disabled or not,
because there is no background color in the checkbox theme.

Instead of adding background color to the edc script, background
images are added because the shape of the check box can be various.

* DefaultTheme/widget/check/check.edc:
* DefaultTheme/widget/check/img_check_bg_enabled.png: Added.
* DefaultTheme/widget/check/img_check_bg_disabled.png: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131576 268f45cc-cd09-0410-ab3c-d52691b4dbfc