floated element with negative margin causes text wrap bug https://bugs.webkit.org/show_bug.cgi?id=94825
Reviewed by Levi Weintraub.

Source/WebCore:

Avoid over-estimating the available width on the line by ensuring that the offset taken to avoid
floats on the line is at least as much as the offset given by border, margin and padding. This only
happens when a negative margin on the float brings its edge back before the offset
given by the border, margin and padding of its parents.

Test: fast/block/float/float-on-line-obeys-container-padding.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::LineWidth::shrinkAvailableWidthForNewFloatIfNeeded):

LayoutTests:

* fast/block/float/float-on-line-obeys-container-padding.html: Added.
* fast/block/float/float-on-line-obeys-container-padding-expected.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131998 268f45cc-cd09-0410-ab3c-d52691b4dbfc