CodeBlock: Give m_putToBaseOperations an inline capacity. <http://webkit.org/b/100190> <rdar://problem/12562466>
Reviewed by Oliver Hunt.

Since the CodeBlock constructor always inserts a single PutToBaseOperation, but there's no
guarantee that more will follow, give the m_putToBaseOperations vector an inline capacity of 1.
There are 4009 of these Vectors on Membuster3, and only 126 of them have more than a single entry.

This change yields a 1.90MB reduction in memory usage.

* bytecode/CodeBlock.h:
(CodeBlock):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132307 268f45cc-cd09-0410-ab3c-d52691b4dbfc