Web Inspector: The bubble for repeated errors is misplaced. https://bugs.webkit.org/show_bug.cgi?id=100525
Reviewed by Pavel Feldman.

The repeated-message bubble is displayed as an inline-block element,
which works well as long as no stack trace is present. If present, the
message is wrapped in an 'ol' element displayed as a block, which pushes
itself down to the next line.

To avoid that issue, this patch switches the wrapper element to flexbox,
glorious flexbox.

* inspector/front-end/inspector.css:
(.console-message .bubble):
(.repeated-message .outline-disclosure):
(.filter-all .console-log-level.repeated-message, .filter-logs .console-log-level.repeated-message):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132772 268f45cc-cd09-0410-ab3c-d52691b4dbfc