Web Inspector: Toolbar overflow appears outside window https://bugs.webkit.org/show_bug.cgi?id=100663
Patch by Patrick Dubroy <dubroy@chromium.org> on 2012-10-29
Reviewed by Pavel Feldman.

Since the search bar was removed from the toolbar, the overflow menu appears too far
to the right, and is unreadable. Fixed this by aligning it relative to the right side
of the window, rather than the left.

* inspector/front-end/Toolbar.js:
(WebInspector.ToolbarDropdown.prototype.show):
* inspector/front-end/inspector.css:
(#toolbar-dropdown .toolbar-label):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132797 268f45cc-cd09-0410-ab3c-d52691b4dbfc