Support captions when PLUGIN_PROXY_FOR_VIDEO https://bugs.webkit.org/show_bug.cgi?id=100690
Reviewed by Simon Fraser.

When built with PLUGIN_PROXY_FOR_VIDEO, WebCore uses a plug-in for the media element's
platform media engine. Update this code path so the shadow DOM elements used to display
text tracks are created and configured correctly.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::configureMediaControls): Create media controls if necessary.

* rendering/RenderEmbeddedObject.cpp:
(WebCore::RenderEmbeddedObject::layout): Set the position and size of the shadow DOM when the
    position of the embedded element changes.
* rendering/RenderEmbeddedObject.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132842 268f45cc-cd09-0410-ab3c-d52691b4dbfc