Block SVG external references pending a security review https://bugs.webkit.org/show_bug.cgi?id=100635
Reviewed by Eric Seidel.

Source/WebCore:

We need to do a security review of loading external SVG references
before we're sure that it is safe.

* css/StyleResolver.cpp:
(WebCore::StyleResolver::createFilterOperations):
* loader/cache/CachedResourceLoader.cpp:
(WebCore::CachedResourceLoader::canRequest):

Source/WTF:

We need to do a security review of loading external SVG references
before we're sure that it is safe.

* wtf/Platform.h:

LayoutTests:

Skip tests that depend on external SVG references.

* platform/chromium/TestExpectations:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132849 268f45cc-cd09-0410-ab3c-d52691b4dbfc