Sparse array size threshold should be increased to 100000 https://bugs.webkit.org/show_bug.cgi?id=100827
Patch by Filip Pizlo <fpizlo@apple.com> on 2012-10-31
Reviewed by Oliver Hunt.

This enables the use of contiguous arrays in programs that previously
couldn't use them. And I so far can't see any examples of this being
a downside. To the extent that there is a downside, it ought to be
addressed by GC: https://bugs.webkit.org/show_bug.cgi?id=100828

* runtime/ArrayConventions.h:
(JSC):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133131 268f45cc-cd09-0410-ab3c-d52691b4dbfc