DOMImplementation should use ScriptWrappable https://bugs.webkit.org/show_bug.cgi?id=101279
Reviewed by Eric Seidel.

PerformanceTests:

This simple PerformanceTest measure how quickly we can access
document.implementation. The goal of this performance test is just to
validate that using ScriptWrappable actually makes the bindings faster.

* Bindings/document-implementation.html: Added.

Source/WebCore:

This patch makes DOMImplementation ScriptWrappable. Both the V8 and JSC
bindings automatically detect the presence of this base class and use
it to optimize getting and setting JavaScript wrappers.

This patch is a 23% progression on Bindings/document-implementation for
the V8 bindings and a 69% progression for the JSC bindings.

* dom/DOMImplementation.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133657 268f45cc-cd09-0410-ab3c-d52691b4dbfc