[CSSRegions] Region styling properties are not filtered correctly https://bugs.webkit.org/show_bug.cgi?id=101768
Reviewed by Andreas Kling.

Source/WebCore:

Region styling allows only a handful of css properties. There is a mechanism in place,
in StyleResolver::isValidRegionStyleProperty that filters the supported region style properties.
In order to be effective, this mechanism relies on RuleData::m_isInRegionRule being set correctly.
This patch makes sure that the flag RuleData::m_isInRegionRule is set properly in RuleData constructor.

Test: fast/regions/region-style-not-supported-properties.html

* css/RuleSet.cpp:
(WebCore::RuleData::RuleData):

LayoutTests:

Visibility is not a supported region style property, therefore using it inside @-webkit-region
should not be possible.

* fast/regions/region-style-not-supported-properties-expected.html: Added.
* fast/regions/region-style-not-supported-properties.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@134085 268f45cc-cd09-0410-ab3c-d52691b4dbfc