[EFL][WK2][AC] Fixed layers are shaking when scrolling https://bugs.webkit.org/show_bug.cgi?id=101549
Patch by Hurnjoo Lee <hurnjoo.lee@samsung.com> on 2012-11-09
Reviewed by Kenneth Rohde Christiansen.

At the time of adjusting positions for fixed layers, the locations of
fixed layers are calculated incorrectly because the contentsSize is empty.

* UIProcess/efl/PageViewportControllerClientEfl.cpp:
(WebKit::PageViewportControllerClientEfl::didChangeContentsSize): Call the
setContentsSize function in LayerTreeCoordinatorProxy in order to calculate
the locations of fixed layers correctly.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@134098 268f45cc-cd09-0410-ab3c-d52691b4dbfc