[V8] HTMLDocument::all() needs not to be custom https://bugs.webkit.org/show_bug.cgi?id=101875
Reviewed by Adam Barth.

We can easily remove the custom getter from V8 because the
custom setter sets .all by ForceSet(). On the other hand,
it is non-trivial to remove the custom getter from JSC
because the custom setter sets .all by putDirect()
and the custom getter wants to get it by getDirect().

No tests. No change in behavior.

* bindings/v8/custom/V8HTMLDocumentCustom.cpp:
* html/HTMLDocument.idl:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@134186 268f45cc-cd09-0410-ab3c-d52691b4dbfc