nrwt: remove a bunch of broken chromium-specific flags https://bugs.webkit.org/show_bug.cgi?id=101979
Reviewed by Tony Chang.

There were a bunch of chromium-specific flags that used to be
supported by NRWT (and handled and passed on to DRT) but have
probably been broken for a long time and are currently ignored.

This patch removes all of those flags; the rule of thumb going forward
is that we should only add flags that affect how the python code functions.
If a flag is just passed verbatim to DRT, the user can use --additional-drt-flag
for that.

* Scripts/webkitpy/layout_tests/run_webkit_tests.py:
(parse_args):
* Scripts/webkitpy/layout_tests/run_webkit_tests_integrationtest.py:
(MainTest.setUp):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@134308 268f45cc-cd09-0410-ab3c-d52691b4dbfc